refactoring of timing
Everything is now based on elapsed time, i.e., the position of the dot is based on how much time has elapsed and the triggering of the synthesizer is based on how much time has elapsed. Seems much more robust in light of processing. Also, in the stop method, I clear the interval and recreate it in start so as to not leave a background loop running.