Fixes to pull request #389
1. As Walter Bender said, it seems very hard to make the canvas update
any faster - it would mean rewriting much of the turtle drawing code. I
think this would probably be best addressed in a separate issue.
2. I decided that adding a separate ‘blink’ over the turtle didn’t look
that nice, but I agree that the current state of affairs could mean
that it is hard to discern which turtle is blinking. I decided to go
for a compromise - the turtle will enlarge slightly during each blink.
3. I have made the longer notes blink for longer than the shorter notes
- specifically each blink will blink for half of the duration of its
respective note. (@Devin, I decided against 25% as with the delays as
they are it’s a little short)