More fixes to Turtle Blink task
I have implemented a note volume scale factor which scales the turtles
according to volume of notes - volume of 1 = 0.804*original, volume of
50 = 1*original, volume of 100 = 1.196*original where original is the
current turtle size * the scale factor of 60/55. You will notice that
this.blinkFinished and this.beforeBlinkSize have been added; these are
due to the fact that sometimes (i.e. when the tab is not in focus)
notes may play before the duration of the previous note is finished. So
if the blink function is called before the previous blink is finished
it will use a cached value of the turtle’s original size to prevent it
using the expanded size as the original size and expanding itself
further. I have also added tweening meaning the turtle will gradually
return to its original size making the exit from the blink more
elegant. However, for this to be possible (and for the delay to be
removed) I had to comment out myTurtle.drawingCanvas.tickEnabled =
false; in js/turtle.js - I hope this will not be an issue as the
comments state you only did this to “prevent some unnecessary refresh
of the canvas”.