Fixed issue #428
A ‘Share Link’ feature has been implemented for the Planet projects.
Also, Planet projects are added by editing the DOM rather than
rewriting the HTML of the PlanetView.

However, when the user publishes a project from local storage, it will
not immediately appear in Planet - the user will have to reload Planet
for the changes to take effect.