Fixed issue 867 (#918)
* Fixed issue 867

eohomegrownapps - GCI
Fixed issue #867

I added support for long and extra long click to the right menu buttons as in the top buttons. I added a merge function when the load button is long-pressed; this skips the sendAllToTrash function when loading a project. I made an icon for long-press of the load button; this is a folder with a merge-ish symbol in the middle.

* Small fix for non-ES6 compatibility

As in title