Fixes issue 919
eohomegrownapps - GCI
Fixes issue 919

I added support for the merge function in Planet - adding a new icon to the thumbnails and adding a merge feature as per pull request #918. I also added merge functionality on extra-long-click (in case a user holds for too long) and for both cases of the `if (_THIS_IS_MUSIC_BLOCKS)` statement.