Fixes issue 866
eohomegrownapps-GCI
Fixes issue #866

1. I added a new icon for saving as WAV - header-icons/save-wav.svg
I have altered the SAVEBOXSVG and a number of other 'magic numbers' in the code to allow for another icon in the box. This process was rather unpleasant - I'll modify these numbers soon to avoid the use of 'magic numbers'.
3. I'm using the FOSS library https://github.com/mattdiamond/Recorderjs (MIT licence) and have included it as a script in HTML - is there anything else I need to do to make it play nicely with require.js?
4. I added a recorder instance and download function to musicutils.js. Similarly to compile, pressing the 'save as WAV' icon starts the code playing with a recording flag - this tells logo.js to start the recorder before code starts and stop it after code starts. The recorder has a built-in 'export as WAV blob' function; this allowed for easy processing and downloading of the WAV blob.

Please do let me know if there are any problems.