Fixes issue #956
This issue was caused by code in activity.css (meant to style the search bar) having too wide a scope and styling all input fields like the search bar - even the fields for names in Planet. I fixed this issue by changing the CSS from `input` to `#search` thus restricting the effect of the styling to only elements with id 'search' (i.e. the search bar).