Record function now uses compile
When the user records a project, MB calls the compile function with a 'recording' argument. This then calls the playback function which records the music as normal. This fixes the first part of @pikurasa's suggested changes - I might sort out the second part tomorrow / at a later date if I have time.