Fixes issue #548
eohomegrownapps-GCI

This fix is in two parts: the addition of a flag selection mechanism to the Planet and the addition of more flags to activity.js.

Flag Selection Mechanism:
I created a div, made visible when the 'Advanced' button is clicked, which displays a list of checkboxes for different flags. When a checkbox is checked, the URL updates automatically. I also fixed a bug with the flag selection tooltips - previously, the flag selection tooltips for projects on the rightmost and leftmost columns would be only partially visible.

Addition of more flags:
As per the task description, I added 'show code blocks', 'collapse code blocks' and 'run code blocks' flags - does exactly what it says on the tin.

Please do let me know if there are any issues, particularly regarding the expected behaviour of different flags - this was complicated by the fact that running the code hides the blocks automatically.