Added screen position blocks to Music Blocks
eohomegrownapps-GCI

As per the task description, I added 'leftpos', 'toppos', 'rightpos', 'bottompos', 'width' and 'height' blocks to the 'media' palette. Please do let me know if the coordinates are meant to be those of the canvas (as in the pull request) or those of the dotted orange 'bounding box' within the canvas.