Fixes issue #923
eohomegrownapps-GCI

Initially, I added a long press function to the stop button ('hard stop'), with the same icon as the normal stop button, which would, in addition to stopping the turtle, set the master volume to 0. (This would not affect any subsequent replays of note blocks or the whole piece, as each time the piece is played the volume is reset to the default.) However, I realised that 1. the stop button disappears after the piece is ended, and 2. the piece 'ends' once the last note is triggered, not finished. This results in the piece being unstoppable if it ends on a long note and the user does not click 'stop' before the last note starts playing. Although I could have implemented something which called the end function once the last note was finished, such a tool (detect if notes are playing) was not easily available in tone.js, and there were too many different cases / edge cases with different types of note to keep a log of when each note will finish relative to the current time. Thus I decided to implement an 'emergency stop' button instead. This button, occupying the same space as the normal stop button, would become visible when the main stop button becomes invisible (and vice versa), and its only purpose is to mute tone.js' master volume and hence stop any stray notes that keep playing after the piece 'ends'. This button has the same icon as the normal stop button, but with a blue background instead of white (because of the way the SVG was designed it was not straightforward to make the white transparent), and with grey instead of read, to signify an 'emergency'/last-resort stop button that shouldn't ordinarily be used.

Sorry about the number of 'hacks' in this PR; please do let me know if any of them will prove a problem for the overall stability of the code so I can make changes accordingly. As always, please do ask if anything is unclear.