Work-around for problem loading from planet with file=
The GlobalPlanet code assumes that the project ID being loaded matches
one of the card that has been cached. But you can load from an id
passed through the URL via file=ID and there is no guarentee that the
project will be one of the cached projects. This really needs a new
approach to how we handle the cache, but we can manage by ignoring the
cache and just loading the downloaded data.

The GlobalPlanet code has also been reformated to match the style of
the other code in the project.