Fixed issue #1221 - chrome security error
The search bar did not appear on some versions of Chromium due to a security policy preventing the execution of inline JS (index.html:46 Refused to execute inline event handler because it violates the following Content Security Policy directive: "script-src 'self' blob: filesystem: chrome-extension-resource:". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution.)

I moved a few instances of inline JS in index.html into the helper.js file, which should (hopefully) solve the problem.

(nb. I have been unable to reproduce the bug so cannot verify that it is fixed myself, but MB still appears to work)