Fixes issue #1245
By using requestAnimationFrame instead of setTimeout, Firefox is forced to wait until the next UI cycle to execute the code meaning it waits for the user to select an option from the dialog before saving. Using setTimeout in the saveHTMLNoPrompt function ensures that if the user is leaving the page (e.g. when reloading) the unload event has time to fire and cancel the download before the download takes place.