Add Loading Animation for Music Blocks (#1988)
* Add container for load animation

I also linked the JSON file that contains the random messages.

* Add CSS for load animation

* Add loading page handler

Run the loading animation when opening projects

* Remove loading page when project finished loading

* Add load messages JSON

This contain the messages for the loading page

* Add Loading GIF

* Add localization for load messages

* Add Spanish Translation for Load Messages

This is just to demo the localization of the loading messages. I only translated it using Google translate so I don't know if the translation is correct.